Add better support for awaiting fields that rely on javascript
The slug preview and upload fields rely on js to render as expected. The addition of the slug preview script, along with better coverage from the upload field script, renamed from "organogram", ensures that puppeteer doesn't beat js to the rendering punch and tests against exactly what the user sees.